+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - The `mapObservable` and `selectProperty` now invoke the mapper even if the source observable value holds a `null` value. The corresponding methods in `ObservableOptionalValue` allow for a more convenient `null` handling.
 - Renamed subscribe methods that accept listeners to `listen`, which invoke the given listener every time the value changes. In contrast, the `subscribe` method also invokes the given subscriber upon registration. 
 - Renamed `EasyBind.listBind` to `bindContent` in order to align with the naming of `Bindings.bindContent`.
+- Moved `PropertyBinding` to `com.tobiasdiez.easybind`.
 
 ### Removed
 - Removed `EasyBind.filter(ObservableValue<T> source, Predicate<? super T> predicate)`. Use `EasyBind.wrapNullable(source).filter(predicate)` instead.